Move the export profile to input data
PWPA-2500
