+++ src/alfasim_score/converter/alfacase/score_input_data.py
@@ -3,13 +3,16 @@ from typing import Dict
 from typing import List
 from typing import Union
 
+import csv
 from barril.units import Scalar
+from pathlib import Path
 
 from alfasim_score.common import AnnulusLabel
 from alfasim_score.common import LiftMethod
 from alfasim_score.constants import ANNULUS_DEPTH_TOLERANCE
 from alfasim_score.constants import FLUID_DEFAULT_NAME
 from alfasim_score.converter.alfacase.score_input_reader import ScoreInputReader
+from alfasim_score.units import LENGTH_UNIT
 from alfasim_score.units import SPECIFIC_HEAT_UNIT
 from alfasim_score.units import THERMAL_CONDUCTIVITY_UNIT
 from alfasim_score.units import THERMAL_EXPANSION_UNIT
@@ -94,3 +97,24 @@ class ScoreInputData:
     def get_default_packer_fluid(self) -> List[Dict[str, Union[Scalar, str]]]:
         """Get the properties of default fluid above packer."""
         return [self._get_default_fluid(FLUID_DEFAULT_NAME)]
+
+    def export_profile_curve(self, filepath: Path, curve_name: str) -> None:
+        """
+        Export the result of a curve to a file.
+        This function export the measured depth related to the well start positions.
+        The exported output file is used to check cases results.
+        """
+        curves = self.reader.read_output_curves()
+        if len(curves):
+            general_data = self.reader.read_general_data()
+            start_position = general_data["water_depth"] + general_data["air_gap"]
+            with open(filepath, "w", encoding="utf-8") as csv_file:
+                writer = csv.DictWriter(csv_file, fieldnames=["measured_depth", curve_name])
+                writer.writeheader()
+                for md, value in zip(curves["measured_depth"], curves[curve_name]):
+                    writer.writerow(
+                        {
+                            "measured_depth": md - start_position.GetValue(LENGTH_UNIT),
+                            curve_name: value,
+                        }
+                    )
